requirements.txt
numpy==2.2.6
librosa==0.11.0
soundfile==0.13.1
requests==2.32.4
pillow==11.2.1
